Add support for inquirer 5
