+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "inquirer-command-prompt",
-  "version": "0.0.6",
+  "version": "0.0.7",
   "license": "MIT",
   "scripts": {
     "lint": "eslint -c .eslintrc 'src/*.js' 'test/*.js'",
@@ -26,16 +26,12 @@
     "bluebird": "^3.5.0",
     "chalk": "^1.1.3",
     "ellipsize": "0.0.3",
-    "inquirer": "^1.1.0",
+    "inquirer": "^5.0.0",
     "preferences": "^0.2.1"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-eslint": "^7.2.3",
-    "babel-preset-es2015": "^6.24.1",
     "chai": "^3.4.1",
     "eslint": "^1.10.3",
-    "istanbul": "^0.4.5",
     "mocha": "^2.3.4",
     "sinon": "^3.2.1"
   }
